backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
redis==5.1.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
python-multipart==0.0.12
sse-starlette==2.1.3
structlog==24.4.0
tenacity==9.0.0
orjson==3.10.7
paho-mqtt==2.1.0
